Refactor events API for consistency
All of the other OCR objects use explicit static Create() methods to
create new instances of an OCR object. The constructors are only used
internally, or implicitly (e.g., when a variable is declared but not
initialized to an OCR object, the default constructor is called, which
typically creates a null object as a placeholder).